+++ crates/myso-crypto/src/ed25519.rs
@@ -268,8 +268,8 @@ impl Verifier<UserSignature> for Ed25519Verifier {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::SuiSigner;
-    use crate::SuiVerifier;
+    use crate::MySoSigner;
+    use crate::MySoVerifier;
     use myso_sdk_types::PersonalMessage;
     use test_strategy::proptest;