+++ crates/myso-crypto/src/secp256r1.rs
@@ -270,8 +270,8 @@ impl Verifier<UserSignature> for Secp256r1Verifier {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::SuiSigner;
-    use crate::SuiVerifier;
+    use crate::MySoSigner;
+    use crate::MySoVerifier;
     use myso_sdk_types::PersonalMessage;
     use test_strategy::proptest;