+++ crates/myso-crypto/src/zklogin/tests.rs
@@ -2,7 +2,7 @@ use signature::Signer;
 use myso_sdk_types::PersonalMessage;
 use myso_sdk_types::ZkLoginInputs;
 
-use crate::SuiVerifier;
+use crate::MySoVerifier;
 use crate::ed25519::Ed25519PrivateKey;
 
 use super::*;