+++ crates/myso-rpc/src/headers.rs
@@ -1,11 +1,11 @@
 /// Chain ID of the current chain
-pub const X_SUI_CHAIN_ID: &str = "x-myso-chain-id";
+pub const X_MYSO_CHAIN_ID: &str = "x-myso-chain-id";
 
 /// Chain name of the current chain
-pub const X_SUI_CHAIN: &str = "x-myso-chain";
+pub const X_MYSO_CHAIN: &str = "x-myso-chain";
 
 /// Current checkpoint height
-pub const X_SUI_CHECKPOINT_HEIGHT: &str = "x-myso-checkpoint-height";
+pub const X_MYSO_CHECKPOINT_HEIGHT: &str = "x-myso-checkpoint-height";
 
 /// Lowest available checkpoint for which transaction and checkpoint data can be requested.
 ///
@@ -14,22 +14,22 @@ pub const X_SUI_CHECKPOINT_HEIGHT: &str = "x-myso-checkpoint-height";
 ///  - transactions
 ///  - effects
 ///  - events
-pub const X_SUI_LOWEST_AVAILABLE_CHECKPOINT: &str = "x-myso-lowest-available-checkpoint";
+pub const X_MYSO_LOWEST_AVAILABLE_CHECKPOINT: &str = "x-myso-lowest-available-checkpoint";
 
 /// Lowest available checkpoint for which object data can be requested.
 ///
 /// Specifically this is the lowest checkpoint for which input/output object data will be
 /// available.
-pub const X_SUI_LOWEST_AVAILABLE_CHECKPOINT_OBJECTS: &str =
+pub const X_MYSO_LOWEST_AVAILABLE_CHECKPOINT_OBJECTS: &str =
     "x-myso-lowest-available-checkpoint-objects";
 
 /// Current epoch of the chain
-pub const X_SUI_EPOCH: &str = "x-myso-epoch";
+pub const X_MYSO_EPOCH: &str = "x-myso-epoch";
 
 /// Current timestamp of the chain - represented as number of milliseconds from the Unix epoch
-pub const X_SUI_TIMESTAMP_MS: &str = "x-myso-timestamp-ms";
+pub const X_MYSO_TIMESTAMP_MS: &str = "x-myso-timestamp-ms";
 
 /// Current timestamp of the chain - encoded in the [RFC 3339] format.
 ///
 /// [RFC 3339]: https://www.ietf.org/rfc/rfc3339.txt
-pub const X_SUI_TIMESTAMP: &str = "x-myso-timestamp";
+pub const X_MYSO_TIMESTAMP: &str = "x-myso-timestamp";